Fixed my typo - read desc
I missed an 's' in subscription - my Subscriber class had a
subcription_price and a subcription_level. Silly mistake but I didn't
catch it for the longest time because I was tab completing (and atom
knew the properties of my class).
It was only once I realized "hey, name is defined, but price isn't" that
I carefully checked the  names of my properties.